+++ Moonlimit.DroneAPI/Moonlimit.DroneAPI.Domain/Service/Generic/IServiceAsync.cs
@@ -9,10 +9,10 @@ namespace Moonlimit.DroneAPI.Domain.Service
     public interface IServiceAsync<Tv, Te>
     {
         Task<IEnumerable<Tv>> GetAll();
-        Task<int> Add(Tv obj);
-        Task<int> Update(Tv obj);
-        Task<int> Remove(int id);
-        Task<Tv> GetOne(int id);
+        Task<Int64> Add(Tv obj);
+        Task<Int64> Update(Tv obj);
+        Task<Int64> Remove(Int64 id);
+        Task<Tv> GetOne(Int64 id);
         Task<IEnumerable<Tv>> Get(Expression<Func<Te, bool>> predicate);
     }
