+++ Moonlimit.DroneAPI/Moonlimit.DroneAPI.Entity/Entity/DroneCom/ObjectDetection.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using NetTopologySuite.Geometries;
@@ -7,8 +8,8 @@ namespace Moonlimit.DroneAPI.Entity.DroneCom
     [Table("object_detections", Schema = "DroneCom")]
     public class ObjectDetection : BaseEntity
     {
-        public int DroneId { get; set; }
-        public int MissionId { get; set; }
+        public Int64 DroneId { get; set; }
+        public Int64 MissionId { get; set; }
         [StringLength(255)]
         public string TagNumber { get; set; }
         [StringLength(255)]
